+++ android/app/src/main/java/com/zzang/chongdae/data/remote/api/OfferingsApiService.kt
@@ -1,8 +1,11 @@
 package com.zzang.chongdae.data.remote.api
 
+import com.zzang.chongdae.data.remote.dto.response.MeetingsResponse
+import com.zzang.chongdae.data.remote.dto.response.OfferingDetailResponse
 import com.zzang.chongdae.data.remote.dto.response.OfferingsResponse
 import retrofit2.Response
 import retrofit2.http.GET
+import retrofit2.http.Path
 import retrofit2.http.Query
 
 interface OfferingsApiService {
@@ -11,4 +14,15 @@ interface OfferingsApiService {
         @Query("last-id") lastOfferingId: Long,
         @Query("page-size") pageSize: Int,
     ): Response<OfferingsResponse>
+
+    @GET("/offerings/{offering-id}")
+    suspend fun getOfferingDetail(
+        @Path("offering-id") offeringId: Long,
+        @Query("member-id") memberId: Long,
+    ): Response<OfferingDetailResponse>
+
+    @GET("/offerings/{offering-id}/meetings")
+    suspend fun getMeetings(
+        @Path("offering-id") offeringId: Long,
+    ): Response<MeetingsResponse>
 }
