+++ android/app/src/main/java/com/zzang/chongdae/data/remote/source/impl/CommentDetailDataSourceImpl.kt
@@ -1,6 +1,7 @@
 package com.zzang.chongdae.data.remote.source.impl
 
-import com.zzang.chongdae.data.remote.api.CommentDetailApiService
+import com.zzang.chongdae.data.remote.api.CommentsApiService
+import com.zzang.chongdae.data.remote.api.OfferingsApiService
 import com.zzang.chongdae.data.remote.dto.request.CommentRequest
 import com.zzang.chongdae.data.remote.dto.response.CommentsResponse
 import com.zzang.chongdae.data.remote.dto.response.MeetingsResponse
@@ -8,11 +9,12 @@ import com.zzang.chongdae.data.remote.source.CommentDetailDataSource
 import retrofit2.Response
 
 class CommentDetailDataSourceImpl(
-    private val service: CommentDetailApiService,
+    private val offeringsApiService: OfferingsApiService,
+    private val commentsApiService: CommentsApiService,
 ) : CommentDetailDataSource {
     override suspend fun getMeetings(offeringId: Long): Result<MeetingsResponse> {
         return runCatching {
-            val response: Response<MeetingsResponse> = service.getMeetings(offeringId)
+            val response: Response<MeetingsResponse> = offeringsApiService.getMeetings(offeringId)
             if (response.isSuccessful) {
                 response.body() ?: error("에러 발생: null")
             } else {
@@ -23,7 +25,7 @@ class CommentDetailDataSourceImpl(
 
     override suspend fun saveComment(commentRequest: CommentRequest): Result<Unit> =
         runCatching {
-            service.postComment(commentRequest = commentRequest)
+            commentsApiService.postComment(commentRequest = commentRequest)
                 .body() ?: throw IllegalStateException()
         }
 
@@ -32,7 +34,8 @@ class CommentDetailDataSourceImpl(
         memberId: Long,
     ): Result<CommentsResponse> =
         runCatching {
-            val response: Response<CommentsResponse> = service.getComments(offeringId, memberId)
+            val response: Response<CommentsResponse> =
+                commentsApiService.getComments(offeringId, memberId)
             if (response.isSuccessful) {
                 response.body() ?: error("에러 발생: null")
             } else {
