+++ android/app/src/main/java/com/zzang/chongdae/data/remote/source/impl/CommentRoomsDataSourceImpl.kt
@@ -1,15 +1,15 @@
 package com.zzang.chongdae.data.remote.source.impl
 
-import com.zzang.chongdae.data.remote.api.CommentRoomsApiService
+import com.zzang.chongdae.data.remote.api.CommentsApiService
 import com.zzang.chongdae.data.remote.dto.response.CommentRoomsResponse
 import com.zzang.chongdae.data.remote.source.CommentRoomsDataSource
 
 class CommentRoomsDataSourceImpl(
-    private val service: CommentRoomsApiService,
+    private val commentsApiService: CommentsApiService,
 ) : CommentRoomsDataSource {
     override suspend fun fetchCommentRooms(memberId: Long): Result<CommentRoomsResponse> {
         return runCatching {
-            service.getCommentRooms(memberId).body() ?: throw IllegalStateException()
+            commentsApiService.getCommentRooms(memberId).body() ?: throw IllegalStateException()
         }
     }
 }
