+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/comment/CommentRoomsFragment.kt
@@ -26,7 +26,7 @@ class CommentRoomsFragment : Fragment(), OnCommentRoomClickListener {
     private val viewModel by viewModels<CommentRoomsViewModel> {
         CommentRoomsViewModel.getFactory(
             CommentRoomsRepositoryImpl(
-                CommentRoomsDataSourceImpl(NetworkManager.commentRoomService()),
+                CommentRoomsDataSourceImpl(NetworkManager.commentsService()),
             ),
         )
     }
