+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/commentdetail/CommentDetailActivity.kt
@@ -24,7 +24,10 @@ class CommentDetailActivity : AppCompatActivity() {
             offeringId,
             offeringTitle,
             CommentDetailRepositoryImpl(
-                CommentDetailDataSourceImpl(NetworkManager.commentDetailService()),
+                CommentDetailDataSourceImpl(
+                    NetworkManager.offeringsService(),
+                    NetworkManager.commentsService(),
+                ),
             ),
         )
     }
